+++ src/retrievals/models/base.py
@@ -15,7 +15,6 @@ class Base(ABC, torch.nn.Module):
         self,
         model: Optional[nn.Module] = None,
         tokenizer: Optional[PreTrainedTokenizer] = None,
-        pooling_method: str = 'cls',
         **kwargs,
     ):
         super().__init__()
@@ -23,12 +22,11 @@ class Base(ABC, torch.nn.Module):
             assert ValueError("Please use AutoModelForEmbedding.from_pretrained(model_name_or_path)")
         self.model = model
         self.tokenizer = tokenizer
-        self.pooling_method = pooling_method
 
     @abstractmethod
     def forward(self, *args, **kwargs):
         """Pytorch forward method."""
-        pass
+        raise NotImplementedError
 
     @abstractmethod
     def encode(self, *args, **kwargs):
@@ -44,6 +42,7 @@ class Base(ABC, torch.nn.Module):
         show_progress_bar: bool = None,
         **kwargs,
     ) -> Union[List[torch.Tensor], np.ndarray, torch.Tensor]:
+        """Encode for sentence embedding"""
         device = device or self.device
         self.model.eval()
         self.model.to(device)
@@ -64,3 +63,26 @@ class Base(ABC, torch.nn.Module):
         else:
             all_embeddings = torch.concat(all_embeddings)
         return all_embeddings
+
+    def preprocess(self, batch_sentence_pair, query_max_length, document_max_length):
+        query_list = [item[0] for item in batch_sentence_pair]
+        document_list = [item[1] for item in batch_sentence_pair]
+
+        query_batch_tokens = self.tokenizer(
+            query_list, padding='max_length', truncation=True, max_length=query_max_length, return_tensors='pt'
+        )
+        query_batch_tokens_on_device = {k: v.to(self.device) for k, v in query_batch_tokens.items()}
+        document_batch_tokens = self.tokenizer(
+            document_list, padding='max_length', truncation=True, max_length=document_max_length, return_tensors='pt'
+        )
+        document_batch_tokens_on_device = {k: v.to(self.device) for k, v in document_batch_tokens.items()}
+
+        return {
+            "query_input_ids": query_batch_tokens_on_device['input_ids'],
+            "query_attention_mask": query_batch_tokens_on_device['attention_mask'],
+            "doc_input_ids": document_batch_tokens_on_device['input_ids'],
+            "doc_attention_mask": document_batch_tokens_on_device['attention_mask'],
+        }
+
+    def gradient_checkpointing_enable(self, gradient_checkpointing_kwargs=None):
+        self.model.gradient_checkpointing_enable(gradient_checkpointing_kwargs=gradient_checkpointing_kwargs)
